update files by controllers
