+++ controllers/listContacts.js
@@ -1,0 +1,8 @@
+const { Contact } = require("../models/contact");
+
+const listContacts = async (req, res) => {
+  const result = await Contact.find({}, "-favorite");
+  res.json(result);
+};
+
+module.exports = listContacts;
