+++ controllers/removeContact.js
@@ -1,0 +1,14 @@
+const { Contact } = require("../models/contact");
+const { HttpError } = require("../helpers");
+
+const removeContact = async (req, res) => {
+  const { contactId } = req.params;
+  const result = await Contact.findByIdAndRemove(contactId);
+  console.log(result);
+  if (!result) {
+    throw HttpError(404, "Not found");
+  }
+  res.json({ message: "contact deleted" });
+};
+
+module.exports = removeContact;
